CORE-5608 Prevent nameserver deletion when editing service without nameserver vars
When editService() is called without nameserver variables set (e.g., when updating
contacts or ID protection), the method was still sending an empty ns array to the
Realtime Register API. This caused nameservers to be cleared both remotely and in
Blesta's database.

Fixed by only calling updateDomain with nameserver parameters when at least one
nameserver is provided in the input variables. This prevents accidental deletion
of nameservers when editing other service properties.